A hangman game
A hangman game is a game of guessing whereby you guess a word by the number of dashes given which represent the character of the word generated.  So if you guess right the program tells you you won,  and print out the word you have guessed right and if you loose, the program tells you that you loose and you should try again. The game is in multiple levels. Level 1 is for easy while level 2 is for difficult.